some more work on win32 API's
git-svn-id: https://svn.code.sf.net/p/jdosbox/code/trunk@516 4cb9e868-4783-4b93-94d1-9e8e689819c6

+++ src/jdos/cpu/Paging.java
@@ -591,8 +591,9 @@ void PrintPageInfo(const char* string, PhysPt lin_addr, bool writing, bool prepa
 
             /*PhysPt*/
             int tableEntryAddr = GetPageTableEntryAddr(addr, dir_entry);
-            table_entry.load((int) Memory.phys_readd(tableEntryAddr));
-            if (table_entry.block.p == 0) Log.exit("Undesired situation 2 in page foiler.");
+            table_entry.load(Memory.phys_readd(tableEntryAddr));
+            if (table_entry.block.p == 0)
+                Log.exit("Undesired situation 2 in page foiler.");
 
             // for debugging...
             if (table_entry.block.base != phys_page)
